+++ Editor/UI/DisableWindow.cs
@@ -3,6 +3,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using UnityEditor;
+using UnityEngine;
 
 #endregion
 
@@ -10,10 +11,10 @@ namespace nadena.dev.ndmf.ui
 {
     internal class DisableWindow : EditorWindow
     {
-        [MenuItem("Tools/NDM Framework/Debug Tools/Enable/Disable Plugins", false, 100)]
+        [MenuItem("Tools/NDM Framework/Debug Tools/Enable-Disable Plugins", false, 100)]
         public static void ShowWindow()
         {
-            GetWindow<DisableWindow>("[NDMF] Enable/Disable Plugins");
+            GetWindow<DisableWindow>("[NDMF] Enable-Disable Plugins");
         }
 
         private List<IPluginInternal> _plugins = null!;
@@ -31,13 +32,33 @@ namespace nadena.dev.ndmf.ui
 
         private void OnDisableChanged(string _, bool _1) => Repaint();
 
+        private Vector2 _scrollPos = Vector2.zero;
         void OnGUI()
         {
             EditorGUILayout.HelpBox(
                 "This window allows you to temporarily disable plugins for the current session.\n" +
                 "Changes you make here will be reverted after restarting Unity.",
                 MessageType.None);
+            
+            using (new EditorGUILayout.HorizontalScope())
+            {
+                if (GUILayout.Button("Enable All Plugins"))
+                {
+                    foreach (var plugin in _plugins)
+                    {
+                        PluginDisablePrefs.SetPluginDisabled(plugin.QualifiedName, false);
+                    }
+                }
+                if (GUILayout.Button("Disable All Plugins"))
+                {
+                    foreach (var plugin in _plugins)
+                    {
+                        PluginDisablePrefs.SetPluginDisabled(plugin.QualifiedName, true);
+                    }
+                }
+            }
 
+            _scrollPos = EditorGUILayout.BeginScrollView(_scrollPos);
             foreach (var plugin in _plugins)
             {
                 var disabled = PluginDisablePrefs.IsPluginDisabled(plugin.QualifiedName);
@@ -47,6 +68,7 @@ namespace nadena.dev.ndmf.ui
                     PluginDisablePrefs.SetPluginDisabled(plugin.QualifiedName, newDisabled);
                 }
             }
+            EditorGUILayout.EndScrollView();
         }
     }
 }